+++ agentic_reasoning/__init__.py
@@ -1,0 +1,1 @@
+from .deep_research import DeepResearcher
