+++ api/apps/sdk/dify_retrieval.py
@@ -16,11 +16,11 @@
 from flask import request, jsonify
 
 from api.db import LLMType
-from api.db.services.dialog_service import label_question
 from api.db.services.knowledgebase_service import KnowledgebaseService
 from api.db.services.llm_service import LLMBundle
 from api import settings
 from api.utils.api_utils import validate_request, build_error_result, apikey_required
+from rag.app.tag import label_question
 
 
 @manager.route('/dify/retrieval', methods=['POST'])  # noqa: F821
